Edited data type of read_ and write_

+++ 0x14-file_io/0-read_textfile.c
@@ -9,7 +9,7 @@
 ssize_t read_textfile(const char *filename, size_t letters)
 {
 	int fd;
-	int read_, write_;
+	ssize_t read_, write_;
 	char *buf = malloc(sizeof(char) * letters);
 
 	if (!buf | !filename)
